Rename remote.js to perapi.js
