+++ lib/replication/staticreplication.js
@@ -6,8 +6,8 @@ const http = require('http')
 const https = require('https')
 
 require('./string')
-const remote = require('./remote')
 const consts = require('./constants')
+const perapi = require('./perapi')
 
 const args = process.argv.slice(2)
 let index = 0
@@ -156,7 +156,7 @@ function addUniqueWithoutSuffix(array, path) {
 }
 
 function checkConsistencyFromStagePOV() {
-    remote.grabFullNodesInfo(host, tenantPath)
+    perapi.grabFullNodesInfo(host, tenantPath)
         .then(array => array.filter(x => !filesLive.check.includes(rootDirPath + x)))
         .then(printStagePOVConsistencyInfo)
 }
